Bootstrap the sizes from window size and layout assumptions
The key idea with the size logic has always been that we know what the
size of each element is relative to everything else. Making the sizes
relative like this has meant that in principle we can scale the contents
up and down to fix different screen resolutions very easily.

The difficulty comes with different aspect ratios. I am assuming that
these won't be too dramatic so we can make some assumptions about the
relative layout positions and work from there. That is what the
`getSizes` function does. It understands that a harmonica will have 10
holes and that the relative size of the columns in the grid should be
a certain degree larger than the space between harp face fragments and a
certain degree smaller than the menu label protrusions.

Using this knowledge we can bootstrap the seed size and figure out the
absolute size of each of these elements.

As it stands there are assumptions here which I would like to dispense
with in due course. I don't want to have to assume that there will only
be 10 holes on the harmonica, and similar with the hard coded 7 rows.

Making these things dynamic will not be easy though because it will
effect the relative sizings of the columns / rows to the fragment
boundaries and menu labels.

I've acknowledged that there's room for using the utility which informs
us of the number of columns which is used elsewhere. There is a TODO
burried in this commit. Go and find it.

As it stands though this should work with those assumptions in place.
It's also a little ugly so this could be improved.