Workaround async screen orientation
The screen orientation happens after the app has loaded which means that
the height and width are 'the wrong way round' for this apps
assumptions.

It's an easy workaround. It doesn't work perfectly if the phone has some
chrome at the top or bottom which changes the height, but it only leaves
the resulting sizing a little bit small rather than really really small
as it was before.

A similar approach is already in use in the menu's to make sure the
render off screen in the beginning.