Force rows to overlap a little
this means that the boundary between the two rows is overlaped slightly
meaning that there is no chance of a rendering issue leaving the
background visible at the boundary.

It's a compromise because it means that the gap between the blow and the
blowbend rows is slightly larger than the others but it's such a small
difference at this is such a quick fix that it's a fair trade off.