Return Pitch and Degree from analysis hook
Now that the Pitch object is a bit more complex, it's no longer
sufficient to just use the id here. I won't delete the `thisPitchId`
return object yet though until I know it's not useful.

The Degree object is probably not directly useful yet, but it's good to
maintain some symmetry.