Present both the sharp and the flat where relevant
This is one solution to the problem that chosing whether to present Ab
or G# in a given cell would depend on some information we don't have.
Namely, which scale from this position on this harp is the user thinking
of.

We could inform the notes presented based on the cells which have been
made active, but that would be extremely sophistocated and may be of
limited help.

This seems to be a reasonably solution.

I was worried it would look cluttered and I'm still not certain that it
won't on smaller devices unfortunately.