Add momoization to the HarpFace
HarpGuru maintains some state which helps it know when to rerender the
menu's. Since a change in the state will trigger a full rerender of the
HarpGuru object, the HarpFace was previously getting rerendered during
interactions which only relate to the visibility of the menu's.

This appears to have had an improvement on the HarpCell responsiveness
to interactions too. I suspect this is becuase even during this
interaction, several rerenders were being triggered either on the
HarpCells or further up.

I wonder whether I've just created a mask for a more fundamental issue
with the data model by doing this. I need to raise an issue to remove
all momoization and test for the root causes of multiple rerenders later
on.