Add memoisation to menu like components
It's not clear that there has been a lot of improvement in performance
here.

It's also not clear to me whether this is safe enough to be justified.

There are hooks inside these components which depend on global state. I
am unsure at this point whether the momoised component needs to observer
changes in these things to allow it's children or more importantly it's
hooks to observe them. I'm pretty sure they don't but I'll need to
experiment with that. If it does then there are a lot of opportunities
to overlook subtle dependencies.

The aim of this change is to reduce the number of renders of the
animation values since I think these are quite expensive.