Add colour to the activity legend and remove text
The text would be nice. Especially if it were able to present the
pitches when degrees are displayed and degrees when pitches were
displayed, but the colour is by far the most important component.

It might be that this thing sholdn't be called activity legend, but
rather color reference or something.

I've had to modify the size logic too be able to encompass the legend
too.