Remove root folder expo framework
And add some basic npm package setup.

There are still expo references left in the project dependencies. These
need to be removed if this will be a project for use by non-expo
projects. The first step will be to get this working as an imported
dependency in the expo based project.