Revert build related changes
This project doesn't need to be built at the moment. I *might* want to
publish it as an npm project at some point for esoteric reasons, but
it's not important for our primary goal, and it's not been approached
here in the most methodical way, so I'll save that effort for a time
when I can do it properly.